feat : 크롤링 완료

+++ BE/mmmmBatch/src/main/java/com/beautifulyomin/mmmmbatch/batch/stock/entity/DailyStockChart.java
@@ -39,7 +39,7 @@ public class DailyStockChart {
     private BigDecimal closingPrice;
 
     @Column(nullable = false)
-    private BigInteger tradingVolume;
+    private Long tradingVolume;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "stockCode", insertable = false, updatable = false)
